+++ habitat-baselines/habitat_baselines/rl/hrl/skills/skill.py
@@ -68,6 +68,7 @@ class SkillPolicy(Policy):
                 break
         if not found_grip and not self.should_ignore_grip:
             raise ValueError(f"Could not find grip action in {action_space}")
+
         self._stop_action_idx, _ = find_action_range(
             action_space, "rearrange_stop"
         )
