+++ habitat-lab/habitat/gym/gym_definitions.py
@@ -47,7 +47,7 @@ def _get_env_name(cfg: "DictConfig") -> Optional[str]:
     return cfg["env_task"]
 
 
-def make_gym_from_config(config: "DictConfig") -> gym.Env:
+def make_gym_from_config(config: "DictConfig", dataset=None) -> gym.Env:
     """
     From a habitat-lab or habitat-baseline config, create the associated gym environment.
     """
@@ -58,7 +58,7 @@ def make_gym_from_config(config: "DictConfig") -> gym.Env:
     assert (
         env_class is not None
     ), f"No environment class with name `{env_class_name}` was found, you need to specify a valid one with env_task"
-    return make_env_fn(env_class=env_class, config=config)
+    return make_env_fn(env_class=env_class, config=config, dataset=dataset)
 
 
 def _add_sim_sensor_to_config(
